few edits

+++ src/controllers/job/homeowner/homeownerJobController.js
@@ -11,8 +11,17 @@ const homeownerCreateJob = async (req, res) => {
     try {
         const userId = req.params.userId;
         const jobDetails = req.body;
+        if (!userId) {
+            throw new Error("User not found");
+        };
+        if (!jobDetails) {
+            throw new Error("Job details not found");
+        };
         if (userId && jobDetails) {
             const job = await jobDatabaseService.createJobHomeowner(userId, jobDetails);
+            if (!job) {
+                throw new Error("Job not found");
+            }
             res.status(201).json(job);
         } else {
             throw new Error("User not found");
@@ -31,8 +40,17 @@ const homeownerUpdateJobTitle = async (req, res) => {
     try {
         const jobId = req.params.jobId;
         const jobTitle = req.body.jobTitle;
+        if (!jobId) {
+            throw new Error("Job not found");
+        }
+        if (!jobTitle) {
+            throw new Error("Job title not found");
+        }
         if (jobId && jobTitle) {
             const job = await jobDatabaseService.updateJobTitleById(jobId, jobTitle);
+            if (!job) {
+                throw new Error("Job not found");
+            }
             res.status(201).json(job);
         } else {
             throw new Error("Job not found");
@@ -48,8 +66,17 @@ const homeownerUpdateJobDescription = async (req, res) => {
     try {
         const jobId = req.params.jobId;
         const jobDescription = req.body.jobDescription;
+        if (!jobId) {
+            throw new Error("Job not found");
+        }
+        if (!jobDescription) {
+            throw new Error("Job description not found");
+        }
         if (jobId && jobDescription) {
             const job = await jobDatabaseService.updateJobDescriptionById(jobId, jobDescription);
+            if (!job) {
+                throw new Error("Job not found");
+            }
             res.status(201).json(job);
         } else {
             throw new Error("Job not found");
@@ -65,8 +92,17 @@ const homeownerUpdateJobQuote = async (req, res) => {
     try {
         const jobId = req.params.jobId;
         const jobQuote = req.body.jobQuote;
+        if (!jobId) {
+            throw new Error("Job not found");
+        }
+        if (!jobQuote) {
+            throw new Error("Job quote not found");
+        }
         if (jobId && jobQuote) {
             const job = await jobDatabaseService.updateJobQuoteById(jobId, jobQuote);
+            if (!job) {
+                throw new Error("Job not found");
+            }
             res.status(201).json(job);
         } else {
             throw new Error("Job not found");
@@ -82,8 +118,17 @@ const homeownerUpdateJobUrgency = async (req, res) => {
     try {
         const jobId = req.params.jobId;
         const jobUrgency = req.body.jobUrgency;
+        if (!jobId) {
+            throw new Error("Job not found");
+        }
+        if (!jobUrgency) {
+            throw new Error("Job urgency not found");
+        }
         if (jobId && jobUrgency) {
             const job = await jobDatabaseService.updateJobUrgencyById(jobId, jobUrgency);
+            if (!job) {
+                throw new Error("Job not found");
+            }
             res.status(201).json(job);
         } else {
             throw new Error("Job not found");
@@ -99,8 +144,17 @@ const homeownerUpdateJob = async (req, res) => {
     try {
         const jobId = req.params.jobId;
         const jobDetails = req.body;
+        if (!jobId) {
+            throw new Error("Job not found");
+        }
+        if (!jobDetails) {
+            throw new Error("Job details not found");
+        }
         if (jobId && jobDetails) {
             const job = await jobDatabaseService.updateJobById(jobId, jobDetails);
+            if (!job) {
+                throw new Error("Job not found");
+            }
             res.status(201).json(job);
         } else {
             throw new Error("Job not found");
@@ -115,8 +169,14 @@ const homeownerUpdateJob = async (req, res) => {
 const homeownerGetAllJobs = async (req, res) => {
     try {
         const userId = req.params.userId;
+        if (!userId) {
+            throw new Error("User not found");
+        }
         if (userId) {
             const jobs = await jobDatabaseService.findJobHomeownerByHomeownerId(userId);
+            if (!jobs) {
+                throw new Error("Jobs not found");
+            }
             res.status(201).json(jobs);
         } else {
             throw new Error("User not found");
@@ -131,8 +191,14 @@ const homeownerGetAllJobs = async (req, res) => {
 const homeownerGetJobById = async (req, res) => {
     try {
         const jobId = req.params.jobId;
+        if (!jobId) {
+            throw new Error("Job not found");
+        }
         if (jobId) {
             const job = await jobDatabaseService.findJobHomeownerById(jobId);
+            if (!job) {
+                throw new Error("Job not found")
+            }
             res.status(201).json(job);
         } else {
             throw new Error("Job not found");
@@ -148,9 +214,12 @@ const homeownerGetJobById = async (req, res) => {
 const homeownerDeleteJob = async (req, res) => {
     try {
         const jobId = req.params.jobId;
+        if (!jobId) {
+            throw new Error("Job not found");
+        }
         if (jobId) {
-            const jobId = await jobDatabaseService.deleteJobById(jobId);
-            res.status(201).json(jobId);
+            const returnedId = await jobDatabaseService.deleteJobById(jobId);
+            res.status(201).json({ message: "Job deleted successfully" });
         } else {
             throw new Error("Job not found");
         }
